--- src/components/App/App.module.css
@@ -1,23 +1,0 @@
-@import url('../../styles/index.css');
-
-.app {
-  text-align: center;
-  height: 100%;
-  min-height: 100vh
-}
-
-.container {
-  margin: 0 auto;
-  max-width: 1280px;
-  display: flex;
-  flex-direction: column;
-}
-
-html{
-  height: 100%;
-}
-body{
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-}
