+++ src/components/AppHeader/AppHeader.module.css
@@ -1,30 +1,29 @@
-
 .header {
-    position: relative;
-    background: var(--background);
-  }
-  
-  .headerWrap {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin: 0px -20px;
-    position: relative;
-  }
-  
-  .headerLogo {
-    position: absolute;
-    left: 50%;
-    transform: translateX(-50%);
-    right: 0;
-    z-index: 1;
-    max-width: 290px;
-    width: 100%;
-  }
-  
-  .headerMenuList {
-    display: flex;
-    list-style: none;
-    margin: 0;
-    padding: 0;
-  }
+  position: relative;
+  background: var(--background);
+}
+
+.headerWrap {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 0px -20px;
+  position: relative;
+}
+
+.headerLogo {
+  position: absolute;
+  left: 50%;
+  transform: translateX(-50%);
+  right: 0;
+  z-index: 1;
+  max-width: 290px;
+  width: 100%;
+}
+
+.headerMenuList {
+  display: flex;
+  list-style: none;
+  margin: 0;
+  padding: 0;
+}
