+++ src/components/BurgerIngredientsSection/BurgerIngredientSection.module.css
@@ -1,6 +1,6 @@
 .burgerIngredientSection {
-    display: flex;
-    width: 100%;
-    flex-wrap: wrap;
-    justify-content: space-between;
-  }
+  display: flex;
+  width: 100%;
+  flex-wrap: wrap;
+  justify-content: space-between;
+}
