+++ src/components/Main/Main.jsx
@@ -2,14 +2,7 @@ import React from "react";
 import PropTypes from "prop-types";
 import BurgerIngredients from "../BurgerIngredients/BurgerIngredients";
 import BurgerConstructor from "../BurgerConstructor/BurgerConstructor";
-import {typeOrder, typeIngridient} from '../../types/types'
-
-BurgerIngredients.propTypes = {
-    items: PropTypes.arrayOf(typeIngridient).isRequired,
-  };
-export default function Main(props: any) {
-    const { items, orders} =  props
-    console.log(props)
+export default function Main({ items, orders}) {
   return (
     <div className="container mb-10">
       <h1 className="text text_type_main-large mt-10 mb-5 title">
