+++ src/styles/index.css
@@ -1,20 +1,3 @@
-.app {
-  text-align: center;
-} 
-
-header{
-  flex: 0 0 auto;
-}
-.main-content{
-  flex: 1 0 auto;
-  
-}
-footer{
-  flex: 0 0 auto;
-}
-body {
-  min-height: 100vh
-}
 @font-face {
   font-family: "JetBrains Mono";
   font-weight: 300;
@@ -56,4 +39,57 @@ body {
   font-family: "JetBrains Mono", sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
+}
+
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
+    monospace;
+}
+
+.container {
+  max-width: 1280px;
+  width: 100%;
+  margin: 0 auto;
+  padding: 0px 20px;
+}
+
+.col {
+  max-width: 600px;
+  width: 100%;
+  padding: 0px 20px;
+}
+
+.title {
+  text-align: left;
+  display: block;
+  width: 100%;
+  font-family: "JetBrains Mono" !important;
+  font-weight: 700;
+}
+
+.row {
+  display: flex;
+  flex-wrap: wrap;
+  margin: 0px -20px;
+}
+
+.constructor-element {
+  margin-bottom: 16px;
+}
+
+.customScroll::-webkit-scrollbar {
+  width: 8px;
+}
+
+.customScroll::-webkit-scrollbar-track {
+  background: var(--background-element);
+}
+
+.customScroll::-webkit-scrollbar-thumb {
+  background-color: var(--text-inactive-color);
+  border-radius: 20px;
+}
+
+.flex {
+  display: flex;
 }
